// web-app/src/types/admin.types.ts
export interface Role {
  UID: string
  Code: string
  RoleNameEn: string
  RoleNameOther?: string // Alias name
  ParentRoleUid?: string // Parent role UID  
  OrgUid?: string // Organization UID
  IsPrincipalRole: boolean
  IsDistributorRole: boolean
  IsAdmin: boolean
  IsWebUser: boolean
  IsAppUser: boolean
  WebMenuData?: string // JSON string
  MobileMenuData?: string // JSON string
  IsActive: boolean
  HaveWarehouse?: boolean // Warehouse access
  HaveVehicle?: boolean // Vehicle access
  IsForReportsTo?: boolean // Reports to access
  CreatedBy?: string
  CreatedTime?: string
  ModifiedBy?: string
  ModifiedTime?: string
}